+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2] - 2025-10-21
+
+### Fixed
+
+- Fix error when unsupported project is in the working directory.
+
 ## [0.2.1] - 2025-10-15
 
 ### Added